pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpu_profiler"
version = "0.1.0"
description = "Uses GPU profiling tools to automatically detect performance issues"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Côme Eyraud" }
]
# Runtime dependencies
dependencies = [
    "hydra-core>=1.3",
]

[project.scripts]
gpu_profiler = "gpu_profiler.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]